migrating data fetching from server to client
