+++ src/styles/ArtistsProfile.module.scss
@@ -30,3 +30,14 @@
 	height: 20rem;
 	border-radius: 0.5rem;
 }
+
+.postsContainerArtistProfile {
+	flex-grow: 1;
+}
+
+.loaderContainer {
+	width: 100vw;
+	height: 100vh;
+	display: flex;
+	justify-content: center;
+}
