+++ src/styles/ClientLayout.module.scss
@@ -1,9 +1,9 @@
 .clientLayout {
-  margin: 0 auto;
-  margin-top: 6rem;
-  max-width: 1200px;
-  display: flex;
-}
-.sectionClientLayout {
-  min-width: 55rem;
+	margin: 0 auto;
+	margin-top: 6rem;
+	display: flex;
+	flex-direction: column;
+	@media screen and (min-width: 600px) {
+		flex-direction: row;
+	}
 }
