+++ src/styles/ConnectionPage.module.scss
@@ -1,0 +1,10 @@
+.loaderContainer {
+	width: 100vw;
+	height: 100vh;
+	display: flex;
+	justify-content: center;
+}
+
+.connectionsContainer {
+	padding: 1rem;
+}
