+++ src/styles/ConnectionsArtists.module.scss
@@ -1,7 +1,15 @@
 .connectionArtistsContainer {
-  margin: 0 auto;
-  margin-top: 75px;
-  padding: 2rem 3rem;
-  max-width: 1200px;
-  height: 100vh;
+	margin: 0 auto;
+	margin-top: 75px;
+	padding: 2rem 3rem;
+	max-width: 1200px;
+	height: 100vh;
+}
+
+.loaderContainer {
+	width: 100vw;
+	height: 100vh;
+	display: flex;
+	justify-content: center;
+	align-items: center;
 }
