+++ src/styles/Posts.module.scss
@@ -1,27 +1,39 @@
 .posts {
-  max-width: 35rem;
-  margin-bottom: 1rem;
+	max-width: 100%;
+	margin-bottom: 1rem;
+	display: flex;
+	flex-direction: column;
+	gap: 0.5rem;
 }
 .postsImage {
-  border-radius: 0.5rem;
+	border-radius: 0.5rem;
+	object-fit: contain;
+	width: 100%;
+	height: 100%;
 }
 .postsInfo {
-  padding: 0.5rem;
+	padding: 0.5rem;
 }
 .postsInfoTriggers {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
 }
 .postsTriggers {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
 }
 .postFooter {
-  margin: 0;
+	margin: 0;
 }
 
 .postshead {
-  padding: 0.2rem;
+	padding: 0.2rem;
+}
+
+.postsImageContainer {
+	width: 100%;
+	height: 300px;
+	margin-top: 2rem;
 }
